+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for searching end devices within an application.
 - Notification during login informing users of unapproved user accounts.
 - Support maximum EIRP value from frequency plans sub-bands.
+- Support duty-cycle value from frequency plans sub-bands.
 
 ### Changed